Add ShrunkIdleState for CorruptRootController. Further refactoring of enemy states.
