+++ Assets/Scripts/Character/Enemy/CorruptRootController.cs
@@ -4,13 +4,6 @@ using UnityEngine.AI;
 [RequireComponent(typeof(Attunable))]
 public class CorruptRootController : EnemyController
 {
-    [SerializeField]
-    private float idleSize = 0.5f;
-    [SerializeField]
-    private float alertSize = 1.0f;
-    [SerializeField]
-    private float sizeChangeRate = 3f; // Rate mesh changes size on state change.
-
     private Attunable attunable;
     
     public override void Start()
@@ -21,36 +14,17 @@ public class CorruptRootController : EnemyController
         
         attunable.onAttuned += Attune;
 
-        TransformUtility.UpdateLocalScale(transform, idleSize);
+        Idle();
+    }
+
+    protected override void Idle()
+    {
+        stateContext.Transition<ShrunkIdleState>();
     }
 
     protected override void Alert()
     {
         base.Alert();
-        Grow();
-    }
-
-    protected override void Idle()
-    {
-        base.Idle();
-        Shrink();
-    }
-
-    private void Shrink()
-    {
-        GradualAction shrink = new GradualAction(UpdateScale, transform.localScale.y, idleSize, sizeChangeRate);
-        ActionManager.Instance.Add(shrink);
-    }
-
-    private void Grow()
-    {
-        GradualAction grow = new GradualAction(UpdateScale, transform.localScale.y, alertSize, sizeChangeRate);
-        ActionManager.Instance.Add(grow);
-    }
-
-    private void UpdateScale(float newScale)
-    {
-        TransformUtility.UpdateLocalScale(transform, newScale);
     }
     
     private void Attune()
