+++ Assets/Scripts/Character/Enemy/EnemyStates/CharacterState.cs
@@ -1,7 +1,9 @@
 using UnityEngine;
 
-public class CharacterState : MonoBehaviour, IState<EnemyController>
+public class CharacterState<T> : MonoBehaviour, IState<T> where T : CharacterController
 {
+    protected T controller;
+
     public virtual void Destroy()
     {
         Destroy(this);
