+++ Assets/Scripts/Character/Enemy/EnemyStates/ShrunkIdleState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0a4f630ff4d26de4f965e64f6df61859
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
